requirements.txt
python-telegram-bot==20.6  # Telegram Bot API wrapper
httpx==0.25.2            # Async HTTP client (compatible with telegram-bot)
redis==5.0.1            # Redis client
pydantic==2.7.1         # Data validation & settings
pytest==8.2.0           # Testing framework
flake8==7.0.0           # Linting
black==24.4.0           # Code formatter
pre-commit==3.7.1       # Git hook manager
mypy==1.10.0            # Static type checker
vaderSentiment==3.3.2           # Sentiment analysis 
pytest-cov>=4.0.0  # For coverage reporting in CI
mkdocs>=1.5.0  # For API documentation
mkdocstrings[python]>=0.24.0  # Python docstring parsing
mkdocs-material>=9.0.0  # Material theme for MkDocs
locust>=2.0.0  # For load testing (dev only)
psutil>=5.9.0  # System monitoring for health checks
aiohttp>=3.9.0  # Async HTTP server for webhooks and health checks
sentry-sdk[flask]>=1.40.0  # Error tracking and monitoring 